TVB-27767: store view_model_disk_size to DB after store_view_model to compute the disk_size
