+++ framework_tvb/tvb/core/entities/model/model_operation.py
@@ -252,7 +252,7 @@ class Operation(Base, Exportable):
     user_group = Column(String, default=None)
     range_values = Column(String, default=None)
     estimated_disk_size = Column(Integer)
-    disk_size = Column(Integer, default=0)
+    view_model_disk_size = Column(Integer, default=0)
 
     algorithm = relationship(Algorithm)
     project = relationship(Project, backref=backref('OPERATIONS', order_by=id, cascade="all,delete"))
