+++ framework_tvb/tvb/tests/framework/core/base_testcase.py
@@ -172,23 +172,6 @@ class BaseTestCase(object):
                 shutil.rmtree(folder, ignore_errors=True)
             os.makedirs(folder)
 
-    @staticmethod
-    def compute_recursive_h5_disk_usage(start_path='.'):
-        """
-        Computes the disk usage of all h5 files under the given directory.
-        :param start_path:
-        :return: A tuple of size in kiB and number of files
-        """
-        total_size = 0
-        n_files = 0
-        for dir_path, _, file_names in os.walk(start_path):
-            for f in file_names:
-                if f.endswith('.h5'):
-                    fp = os.path.join(dir_path, f)
-                    total_size += os.path.getsize(fp)
-                    n_files += 1
-        return int(round(total_size / 1024.)), n_files
-
     def count_all_entities(self, entity_type):
         """
         Count all entities of a given type currently stored in DB.
