+++ framework_tvb/tvb/tests/framework/core/services/project_service_test.py
@@ -296,7 +296,7 @@ class TestProjectService(TransactionalTestCase):
             assert '0.0 KiB' != project.disk_size_human
 
             prj_folder = self.structure_helper.get_project_folder(project)
-            actual_disk_size = self.compute_recursive_h5_disk_usage(prj_folder)[0]
+            actual_disk_size = FilesHelper.compute_recursive_h5_disk_usage(prj_folder)[0]
 
             ratio = float(actual_disk_size) / project.disk_size
             msg = "Real disk usage: %s The one recorded in the db : %s" % (actual_disk_size, project.disk_size)
